# contracts/openapi.json
{
  "components": {
    "schemas": {
      "APIErrorDetail": {
        "description": "The stable error payload shared by all public API domains.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "APIErrorDetail",
        "type": "object"
      },
      "APIErrorResponse": {
        "description": "The standard COMPASS JSON error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "APIErrorResponse",
        "type": "object"
      },
...
      "AppointmentPolicy": {
        "enum": [
          "NONE",
          "OPTIONAL",
          "REQUIRED"
        ],
        "title": "AppointmentPolicy",
        "type": "string"
      },
...
      "DeliveryMode": {
        "enum": [
          "IN_PERSON",
          "ONLINE"
        ],
        "title": "DeliveryMode",
        "type": "string"
      },
...
      "ProviderRoleCode": {
        "enum": [
          "COUNSELOR",
          "GUIDANCE_SERVICES_STAFF"
        ],
        "title": "ProviderRoleCode",
        "type": "string"
      },
...
      "ServiceCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "appointment_policy": {
            "$ref": "#/components/schemas/AppointmentPolicy"
          },
          "cancellation_cutoff_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Cutoff Minutes"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "default_duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Duration Minutes"
          },
          "delivery_modes": {
            "items": {
              "$ref": "#/components/schemas/DeliveryMode"
            },
            "title": "Delivery Modes",
            "type": "array"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider_roles": {
            "items": {
              "$ref": "#/components/schemas/ProviderRoleCode"
            },
            "title": "Provider Roles",
            "type": "array"
          }
        },
        "required": [
          "code",
          "name",
          "appointment_policy"
        ],
        "title": "ServiceCreateRequest",
        "type": "object"
      },
      "ServiceListResponse": {
        "additionalProperties": false,
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ServiceResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "has_next"
        ],
        "title": "ServiceListResponse",
        "type": "object"
      },
      "ServiceResponse": {
        "additionalProperties": false,
        "properties": {
          "appointment_policy": {
            "$ref": "#/components/schemas/AppointmentPolicy"
          },
          "cancellation_cutoff_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Cutoff Minutes"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Duration Minutes"
          },
          "delivery_modes": {
            "items": {
              "$ref": "#/components/schemas/DeliveryMode"
            },
            "title": "Delivery Modes",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider_roles": {
            "items": {
              "$ref": "#/components/schemas/ProviderRoleCode"
            },
            "title": "Provider Roles",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "description",
          "appointment_policy",
          "default_duration_minutes",
          "cancellation_cutoff_minutes",
          "delivery_modes",
          "provider_roles",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "ServiceResponse",
        "type": "object"
      },
...
      "ServiceUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "appointment_policy": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AppointmentPolicy"
              }
            ],
            "default": "NONE"
          },
          "cancellation_cutoff_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Cutoff Minutes"
          },
          "default_duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Duration Minutes"
          },
          "delivery_modes": {
            "items": {
              "$ref": "#/components/schemas/DeliveryMode"
            },
            "title": "Delivery Modes",
            "type": "array"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "provider_roles": {
            "items": {
              "$ref": "#/components/schemas/ProviderRoleCode"
            },
            "title": "Provider Roles",
            "type": "array"
          }
        },
        "title": "ServiceUpdateRequest",
        "type": "object"
      },
...
      "ValidationIssue": {
        "description": "A safe, normalized validation issue returned inside the error envelope.",
        "properties": {
          "loc": {
            "items": {
              "type": "string"
            },
            "title": "Loc",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "message",
          "type"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OpaqueSessionAuth": {
        "description": "HttpOnly opaque COMPASS session cookie. The browser sends it automatically; it is not a Bearer token. State-changing requests also require Django CSRF protection.",
        "in": "cookie",
        "name": "compass_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The version-one backend API for COMPASS.",
    "title": "COMPASS API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/services": {
      "get": {
        "operationId": "servicesList",
        "parameters": [
          {
            "in": "query",
            "name": "include_inactive",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Inactive",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "appointment_policy",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AppointmentPolicy"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceListResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "OpaqueSessionAuth": []
          }
        ],
        "summary": "Services List",
        "tags": [
          "services"
        ]
      },
      "post": {
        "operationId": "servicesCreate",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            },
            "description": "Created"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "OpaqueSessionAuth": []
          }
        ],
        "summary": "Services Create",
        "tags": [
          "services"
        ]
      }
    },
    "/api/v1/services/{service_id}": {
      "get": {
        "operationId": "servicesGet",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "OpaqueSessionAuth": []
          }
        ],
        "summary": "Services Get",
        "tags": [
          "services"
        ]
      },
      "patch": {
        "operationId": "servicesUpdate",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "OpaqueSessionAuth": []
          }
        ],
        "summary": "Services Update",
        "tags": [
          "services"
        ]
      }
    },
    "/api/v1/services/{service_id}/disable": {
      "post": {
        "operationId": "servicesDisable",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "OpaqueSessionAuth": []
          }
        ],
        "summary": "Services Disable",
        "tags": [
          "services"
        ]
      }
    },
    "/api/v1/services/{service_id}/enable": {
      "post": {
        "operationId": "servicesEnable",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "OpaqueSessionAuth": []
          }
        ],
        "summary": "Services Enable",
        "tags": [
          "services"
        ]
      }
    }
  },
  "servers": [],
...
}